+++ src/main/java/de/maxhenkel/pipez/ItemFilter.java
@@ -70,6 +70,8 @@ public class ItemFilter extends Filter<Item> {
         if (compound.contains("Destination", Constants.NBT.TAG_COMPOUND)) {
             destination = new DirectionalPosition();
             destination.deserializeNBT(compound.getCompound("Destination"));
+        } else {
+            destination = null;
         }
 
         if (compound.contains("Invert", Constants.NBT.TAG_BYTE)) {